README: correcting begin/end timestamp format
Correcting `begintimestamp` & `endtimestamp` format to match the Regular Expression test in the code.

`imagedirector.delete` kept failing, I was following the 'mm/dd/yyyy hh:mm:ss' format specified in the documentation. It wasn't until I took a screenshot of the error message that I figured out why. The regular expression test was failing because the code wants the day to appear before the month. I have corrected the documentation so that the next person could have better luck.

*Error: "Failed to call service imagedirectory.delete_files. value 02/26/2021 00:00:00 does not match regular expression [0-3][0-9]/[0-1][0-9]/\d{4} [0-2]:[0-5][0-9]:[0-5][0-9] for dictionary value @ data['endtimestamp']*